login: move account creation into services/accounts.py
/open-account's body -- open-registration gate, email/password validation,
audit log, invitation-link consumption, duplicate email/handle checks,
verification-code check and user creation -- moves into a new
services/accounts.py. The dead-but-kept pay_reward_for_invitation helper (no
callers; carried a "should remove out of api" TODO) moves with it, since it is
the same domain: what an invitation pays out once an account opens.

This removes login.py's import of responders.user (a level-1 -> level-3 import
that SS1.2 forbids); the responder is now called from the service.

The function-local `from ... import invitations as invitations_service` inside
the endpoint becomes a normal module-level import in the service -- the cycle
it was dodging only existed because the code was in api/.

No HTTP contract change: same path, method, response_model, Body(...)
parameters (so the generated Body_create_user_open_... schema is unchanged) and
the same get_request_context dependency, which is what puts the
OAuth2PasswordBearer security entry on this operation.
